% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{brute_merge}
\alias{brute_merge}
\title{brute_merge}
\usage{
brute_merge(so1, so2)
}
\arguments{
\item{so1}{is a SpatialPolygonsDataFrame}

\item{so2}{is a SpatialPolygonsDataFrame}
}
\value{
a merged spdf
}
\description{
merge spdfs by adding unique columns to the dataframe
}
