% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{month_no}
\alias{month_no}
\title{month_no}
\usage{
month_no(vec)
}
\arguments{
\item{vec}{is a character vector of month names}
}
\value{
a vector of number 1-12 representing months
}
\description{
converts month names to numbers
}
